Split wiznet.WIZNET5K off into its own file
